requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-pptx==0.6.23
python-multipart==0.0.9